add another event handler for risk profile on new fingerprints
