+++ tslint.json
@@ -13,7 +13,8 @@
             true,
             "never-prefix"
         ],
-        "object-literal-sort-keys": false
+        "object-literal-sort-keys": false,
+        "max-classes-per-file": false
     },
     "rulesDirectory": []
 }
